+++ src/lib.rs
@@ -57,6 +57,7 @@
 //! method. Merge is guaranteed to be commutative, associative, and idempotent.
 
 #![cfg_attr(not(feature = "std"), no_std)]
+#![warn(missing_docs)]
 
 extern crate alloc;